+++ extra/modules/pb-response-correction/src/main/java/org/prebid/server/hooks/modules/pb/response/correction/core/config/model/AppVideoHtmlConfig.java
@@ -1,0 +1,15 @@
+package org.prebid.server.hooks.modules.pb.response.correction.core.config.model;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.Value;
+
+import java.util.List;
+
+@Value(staticConstructor = "of")
+public class AppVideoHtmlConfig {
+
+    boolean enabled;
+
+    @JsonProperty("excluded-bidders")
+    List<String> excludedBidders;
+}
