--- extra/modules/response-correction/pom.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-
-    <parent>
-        <groupId>org.prebid.server.hooks.modules</groupId>
-        <artifactId>all-modules</artifactId>
-        <version>3.12.0-SNAPSHOT</version>
-    </parent>
-
-    <artifactId>response-correction</artifactId>
-
-    <name>response-correction</name>
-    <description>Response correction module</description>
-</project>
