--- extra/modules/response-correction/src/main/java/org/prebid/server/hooks/modules/response/correction/core/correction/Correction.java
@@ -1,11 +1,0 @@
-package org.prebid.server.hooks.modules.response.correction.core.correction;
-
-import org.prebid.server.auction.model.BidderResponse;
-import org.prebid.server.hooks.modules.response.correction.core.config.model.Config;
-
-import java.util.List;
-
-public interface Correction {
-
-    List<BidderResponse> apply(Config config, List<BidderResponse> bidderResponses);
-}
